wuying-agentbay-sdk-main/cookbook/envs/codespace/auto-testing-agent/langchain/requirements.txt
wuying-agentbay-sdk>=0.7.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
python-dotenv>=1.0.0